+++ onchain/bench/Cases/GrowingTreeClaim.hs
@@ -1,3 +1,7 @@
+{-# LANGUAGE ImportQualifiedPost #-}
+{-# LANGUAGE NamedFieldPuns #-}
+{-# LANGUAGE OverloadedStrings #-}
+
 module Cases.GrowingTreeClaim (growingTreeClaim) where
 
 import Bench (Bench, bcfgSigningKeyFilePath)
@@ -32,6 +36,7 @@ import TrustlessSidechain.Types (
   chainId,
   genesisHash,
   genesisUtxo,
+  governanceAuthority,
   mteAmount,
   mteIndex,
   mtePreviousMerkleRoot,
@@ -62,6 +67,7 @@ growingTreeClaim = do
   -- problem.
   signingKeyFile <- Reader.asks bcfgSigningKeyFilePath
   addr <- Bench.readAddr
+  governanceAuthority <- Bench.readGovernanceAuthority
 
   -- Benchmark suite
   --------------------
@@ -90,6 +96,7 @@ growingTreeClaim = do
             , genesisUtxo = txOutRef
             , thresholdNumerator = 2
             , thresholdDenominator = 3
+            , governanceAuthority
             }
 
     -- Iniatialising the sidechain: