+++ onchain/bench/Cases/InitSidechain.hs
@@ -1,3 +1,7 @@
+{-# LANGUAGE ImportQualifiedPost #-}
+{-# LANGUAGE NamedFieldPuns #-}
+{-# LANGUAGE OverloadedStrings #-}
+
 module Cases.InitSidechain (initSidechainBench) where
 
 import Bench (Bench, bcfgSigningKeyFilePath)
@@ -21,6 +25,7 @@ import TrustlessSidechain.Types (
   chainId,
   genesisHash,
   genesisUtxo,
+  governanceAuthority,
   thresholdDenominator,
   thresholdNumerator,
  )
@@ -33,8 +38,8 @@ initSidechainBench = do
 
   signingKeyFile <- Reader.asks bcfgSigningKeyFilePath
 
-  -- TODO: urgh, we really shouldn't do this so fix this later...
-  addr <- IO.Class.liftIO $ readFile "payment.addr"
+  addr <- Bench.readAddr
+  governanceAuthority <- Bench.readGovernanceAuthority
 
   -- Benchmark suite
   --------------------
@@ -64,6 +69,7 @@ initSidechainBench = do
               , genesisUtxo = txOutRef
               , thresholdNumerator = 2
               , thresholdDenominator = 3
+              , governanceAuthority
               }
 
       --  Building the initial committee: