+++ onchain/test/Test/TrustlessSidechain/UpdateCommitteeHashMessage.hs
@@ -4,11 +4,12 @@ module Test.TrustlessSidechain.UpdateCommitteeHashMessage (test) where
 
 import Data.ByteString (ByteString)
 import Data.ByteString.Base16 qualified as Base16
-import Ledger (TxId (TxId), TxOutRef (TxOutRef))
+import Ledger (PubKeyHash (PubKeyHash), TxId (TxId), TxOutRef (TxOutRef))
 import PlutusTx.Builtins (blake2b_256)
 import PlutusTx.Prelude
 import Test.Tasty (TestTree, testGroup)
 import Test.Tasty.HUnit (testCase, (@?=))
+import TrustlessSidechain.Governance (mkGovernanceAuthority)
 import TrustlessSidechain.Types (
   GenesisHash (GenesisHash),
   SidechainParams (SidechainParams),
@@ -25,13 +26,14 @@ unitTests =
   let conv = (toBuiltin @ByteString) . Base16.decodeLenient
       genesisHash = GenesisHash (conv "e8118a6a0f2ea8447b2418b0301fa53fa97f95a042fc92edbd7eda9f809d9040")
       genesisUtxo = TxOutRef (TxId (conv "7247647315d327d4e56fd3fe62d45be1dc3a76a647c910e0048cca8b97c8df3e")) 0
+      governanceAuthority = mkGovernanceAuthority $ PubKeyHash "4f2d6145e1700ad11dc074cad9f4194cc53b0dbab6bd25dfea6c501a"
       pubKeys =
         [ SidechainPubKey (conv "029ef0f8e7f2144461246f4d14772a34945069e0b746fc93d641f5cceb23dba760")
         , SidechainPubKey (conv "0374df74bd17e647fbea5ad07699ee36eae9247b2fb633f31223da66626e083272")
         , SidechainPubKey (conv "038b098f1a3ccb005419df63dc1ce954a3f9071e2f41aefece0f86ee991285b498")
         , SidechainPubKey (conv "03f16df0d21e2a447d820999c6b65794820cd1920cafccc3a7b83956f6148441ed")
         ]
-      scParams = SidechainParams 78 genesisHash genesisUtxo 2 3
+      scParams = SidechainParams 78 genesisHash genesisUtxo 2 3 governanceAuthority
       epoch = 123
       previousMRH = conv "abababababababababababababababababababababababababababababababab"
       withoutPreviousMRH = UpdateCommitteeHashMessage scParams pubKeys Nothing epoch
@@ -39,12 +41,12 @@ unitTests =
 
       actualCbor1 = serialiseUchm withoutPreviousMRH
       actualCbor2 = serialiseUchm withPreviousMRH
-      expectedCbor1 = conv "d8799fd8799f184e5820e8118a6a0f2ea8447b2418b0301fa53fa97f95a042fc92edbd7eda9f809d9040d8799fd8799f58207247647315d327d4e56fd3fe62d45be1dc3a76a647c910e0048cca8b97c8df3eff00ff0203ff9f5821029ef0f8e7f2144461246f4d14772a34945069e0b746fc93d641f5cceb23dba76058210374df74bd17e647fbea5ad07699ee36eae9247b2fb633f31223da66626e0832725821038b098f1a3ccb005419df63dc1ce954a3f9071e2f41aefece0f86ee991285b498582103f16df0d21e2a447d820999c6b65794820cd1920cafccc3a7b83956f6148441edffd87a80187bff"
-      expectedCbor2 = conv "d8799fd8799f184e5820e8118a6a0f2ea8447b2418b0301fa53fa97f95a042fc92edbd7eda9f809d9040d8799fd8799f58207247647315d327d4e56fd3fe62d45be1dc3a76a647c910e0048cca8b97c8df3eff00ff0203ff9f5821029ef0f8e7f2144461246f4d14772a34945069e0b746fc93d641f5cceb23dba76058210374df74bd17e647fbea5ad07699ee36eae9247b2fb633f31223da66626e0832725821038b098f1a3ccb005419df63dc1ce954a3f9071e2f41aefece0f86ee991285b498582103f16df0d21e2a447d820999c6b65794820cd1920cafccc3a7b83956f6148441edffd8799f5820ababababababababababababababababababababababababababababababababff187bff"
+      expectedCbor1 = conv "d8799fd8799f184e5820e8118a6a0f2ea8447b2418b0301fa53fa97f95a042fc92edbd7eda9f809d9040d8799fd8799f58207247647315d327d4e56fd3fe62d45be1dc3a76a647c910e0048cca8b97c8df3eff00ff020358383466326436313435653137303061643131646330373463616439663431393463633533623064626162366264323564666561366335303161ff9f5821029ef0f8e7f2144461246f4d14772a34945069e0b746fc93d641f5cceb23dba76058210374df74bd17e647fbea5ad07699ee36eae9247b2fb633f31223da66626e0832725821038b098f1a3ccb005419df63dc1ce954a3f9071e2f41aefece0f86ee991285b498582103f16df0d21e2a447d820999c6b65794820cd1920cafccc3a7b83956f6148441edffd87a80187bff"
+      expectedCbor2 = conv "d8799fd8799f184e5820e8118a6a0f2ea8447b2418b0301fa53fa97f95a042fc92edbd7eda9f809d9040d8799fd8799f58207247647315d327d4e56fd3fe62d45be1dc3a76a647c910e0048cca8b97c8df3eff00ff020358383466326436313435653137303061643131646330373463616439663431393463633533623064626162366264323564666561366335303161ff9f5821029ef0f8e7f2144461246f4d14772a34945069e0b746fc93d641f5cceb23dba76058210374df74bd17e647fbea5ad07699ee36eae9247b2fb633f31223da66626e0832725821038b098f1a3ccb005419df63dc1ce954a3f9071e2f41aefece0f86ee991285b498582103f16df0d21e2a447d820999c6b65794820cd1920cafccc3a7b83956f6148441edffd8799f5820ababababababababababababababababababababababababababababababababff187bff"
       actualHash1 = blake2b_256 expectedCbor1
       actualHash2 = blake2b_256 expectedCbor2
-      expectedHash1 = conv "29851edc4a500fcb6b74597107c202e2a2d167b497c461a4717b1bb00cb3f70b"
-      expectedHash2 = conv "cee6b4e5cbb9830a82d5b72cc3e8f75dd5ad7d23027b7e16245582048c7cb2b0"
+      expectedHash1 = conv "8a72bf4791b12d3788ce95898aacdd316abfff7db5a98f89ef256b15db8e6048"
+      expectedHash2 = conv "1aeb69dfbe8f5f098424bd8312b9585dbd4c2c131f515d4a3de1278436084e66"
    in testGroup
         "UpdateCommitteeHashMessage"
         [ testCase "cbor of message without previous merkle root hash" $ actualCbor1 @?= expectedCbor1